Learnt about headers, CORS etc

+++ db/db_user.py
@@ -1,3 +1,4 @@
+from fastapi import HTTPException, status
 from sqlalchemy.orm.session import Session
 from db.hash import Hash
 from schemas import UserBase
@@ -20,11 +21,16 @@ def get_all_users(db: Session):
 
 
 def get_one_user(db: Session, id: int):
-    return db.query(DbUser).filter(DbUser.id == id).first()
+    user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"User with id {id} not found")
+    return user
 
 
 def update_user(db: Session, id: int, request: UserBase) -> str:
     user = db.query(DbUser).filter(DbUser.id == id)
+    if not user.first():
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"User with id {id} not found")
     user.update({
                 DbUser.username: request.username,
                 DbUser.email: request.email,
@@ -35,6 +41,8 @@ def update_user(db: Session, id: int, request: UserBase) -> str:
 
 def delete_user(db: Session, id: int):
     user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"User with id {id} not found")
     db.delete(user)
     db.commit()
     return "OK"
